nextcloud-exam/requirements.txt
python-dotenv
Flask
Flask-APScheduler
Flask-Mail